Changed wtosc phase increment field from 16:16 to 8:24
The 16:16 phase increment field wasn't sufficiently accurate
for the lower octaves, which resulted in clearly audible phasing
issues, particularly between low notes at octave intervals.

Note that the phase accumulator is still 24:8 (not 16:16 - the
comment was incorrect), which means we should still be able to
handle waveforms up to 16 Msamples.

Also note that although the phase increment is converted to 24:8
internally to match the phase accumulator, sufficient accuracy
is maintained as a side effect of the waveform period. (2048, or
11 bits for the built-in waves in the low octave.) As these bits
fall off with higher octaves due to the mipmapping, they're
replaced by bits from 'dphase', so that's not an issue either.

Fixes #10.