wtosc now checks the length of selected waves
The wtosc wavetable oscillator unit now checks that it can actually
handle the length of a wave when the 'w' register is written.

(There is currently a limit slightly short of 16 Msamples due to
the 24:8 fixed point phase accumulator, and the loop/stop logic
requiring that the accumulator can run slightly beyond the end of
the wave without wrapping.)

Fixes #31.