Fixed some VM timing issues
Control register write callbacks are now performed BEFORE bumping
the timer with the delay duration, so that voice units can derive
the current subsample time from that.

The wtosc phase register write callback now reverses the timer
value, as it's actually the time REMAINING of the current sample
frame.

Fixes #43.