Slightly more accurate VM timing and wtosc pitch
Added correct rounding in a few places to have slightly improved
accuracy in the P2D, DELAY and TDELAY instructions, and wtosc unit
pitch control.

This is only a slight improvement. We'd have to change timing
units and/or register format, or switch to floating point to fix
this issue properly.

Closes #48 for now, but doesn't really fix it...