Correct subvoice spawn and event timing
Added a new VM instruction INITV that populates the voice as a
program actually starts running. This ensures that voice units are
initialized at the exact time intended, rather than at the start
of the current fragment.

The SPAWN* instructions now take in account that inlined subvoices
(as opposed to independent subvoices) actually do initialize and
start during the very sample frame that the instruction is issued.

The wtosc unit now considers the voice timer when initializing, so
that the oscillator starts at phase 0 exactly when intended, with
subsample accuracy.

Fixes #44.