Adjusted filter12 'q' clamping to avoid explosions
The filter12 unit wasn't stable at high cutoff frequencies when the
resonance is set close to zero. Preferably, it should be possible
to make the filter completely non-resonant, but as it is, that only
works at cutoff frequencies a bit below the Nyqvist/4 limit.

For now, I'm just clamping the 'q' register at 512 instead of 256.

Fixes #53.