Removed the 'run' statement hack
The old RUN instruction just wasn't compatible with how the engine
works now, with the modular voices and all, and was replaced with
a hack where the compiler just issues a SPAWN followed by a WAIT.
Fragment resolution! Not even sample accurate, let alone subsample.
As if that isn't bad enough, it was hardwired to use the top
subvoice slot index.

It turns out that it would be really hard to implement this
properly, so we're just going to drop it, as the current hack is
too inaccurate to be of much use.

The 'wait' statement/instruction is still there if you really want
this feature...

Closes #45.