Added a2_GetUnitDescriptor() API call
This "new" call is a public version of the internal a2_GetUnit()
inline, allowing applications to get hold of A2_unitdesc instances
of registered units.

A2_unitdesc is already a public interface that comes with the unit
API, so nothing new there - and it provides all the information we
need at this point.

Closes #52.