Added support for driver options
a2_NewDriver() now accepts a comma separated list of options
appended after the driver name. These options will be parsed
and copied into the optc/optv array in the driver struct.

Added a2_DestroyDriver(), which wraps the destructor method
and takes care of the options array, if any.

Driver Destroy() method is now optional. Default action is
to free() the struct. Destroy() methods removed from built-in
drivers, as none of them need anything special at this point.

JACK driver takes a non-switch (with no '-' in front of it)
option as a custom client name instead of "Audiality 2".

Closes #82.