Static descriptors for built-in drivers, eliminating leaks
The descriptors for the built-in driver are now in a static
array, so there will be no memory leaks there no matter what!

Applications can still unregister some or all built-in drivers,
register user drivers etc; this is just a more robust
implementation in terms of avoiding irrelevant memory leak
warnings from Valgrind and similar.

Fixes #85.