Added rtsubtest.c
Added a simple test, rtsubtest.c, that loads a song and plays two
instances of it on different realtime states. It sets up a state
with a substate, both realtime with actual audio drivers, so that
the states are asynchronously playing from the same physical song
and wave data.

The purpose of this test is to verify that substates actually work,
and that they can indeed piggyback on the data of their master
states as intended.