Root voice always has two or more channels
As a temporary hack to handle songs and sounds that expect two or
more output channels, we now use a special root voice program for
mono master output, that runs the whole graph in two channel stereo,
and mixes down to mono right before the master xinsert unit.

As it is, we could avoid the need for this by adding one channel
support to the units still missing it (probably fbdelay at this
point), but dealing with manually wired voice structures isn't
going to be quite that easy.