xinsert terminology fixed - again
"Send" and "Return" weren't the best terms, as they made little
sense outside the context of 'xinsert' and the engine internals
surrounding it. A2 "sends" and "returns" would typically NOT be
used as insert effect pairs, but rather separately, and in that
context, this terminology was just weird and confusing.

So, to make more sense in actual use cases, and also match some
popular APIs, the terminology is now as such:

xinsert - sends audio through external code
    a2_InsertCallback()

xsource - streams audio from external code
    a2_SourceCallback()
    a2_OpenSource()

xsink - streams audio to external code
    a2_SinkCallback()
    a2_OpenSink()

Closes #139.