HACK: Fixed clicks in conjunction with voice termination
Sample accurate voice termination (#121) somehow causes clicks
in some cases, suggesting that subvoices aren't always set up
to die at the exact same time as their parent. It's still not
known why exactly this is, but by allowing the subvoices to run
until the end of the originally requested fragment, the problem
is avoided, while still avoiding the pre #121 issue of subvoices
not processing the current fragment at all.

Closes #145.