'import' directive now tries the directory of the current file first
The compiler will now grab the directory path (if specified) of the current
file, and prepend that before import file names, before trying the import
name as is.

This allows a module to find imports using relative paths, regardless of
what the current directory is when the module itself is loaded.

Closes #149.