'import' directive namespace support
The 'import' directive can now be followed by the new 'as' keyword, and
a namespace name. Imported symbols will be put in that namespace instead
of pulled into the root namespace of the current file.

The compiler <object/namespace>.<name> construct can now look up arbitrary
symbols in namespaces (not just control registers of named units), and it
will also look up imports in banks. (The latter is what allows imports to
be placed in namespaces.)

Some compiler refactoring and a bug fix: The compiler no longer looks in
the root import list when asked to look only in a namespace.

Of course, another test/example has been added: importtest2.a2s. This one
imports a2jingle.a2s, and plays it back through a small, crappy AM radio.

Closes #147.