Fixed inplace unsafe process() cases in 'panmix' unit
The 'panmix' unit wasn't inplace safe in some cases. It was reading some
input samples after writing to output buffers, which is of course not safe
when the wiring logic assumes buffers can be reused without restriction.

Restored importest2.a2s to the orinigal master voice structure, which was
how this issue was discovered in the first place.

Closes #166.