Implemented '@' shorthand for "<reg> <expr>; set <reg>"
It is now possible to prefix a <reg> <expression> statement with '@' to
have the compiler issue a SET instruction for <reg> after calculating its
final value.

The register allocator now keeps track of what type of use each register
was allocated for, and a2_Code() will throw an error if SET is used on
a register that isn't mapped to a voice or unit control register.

"fmtest.a2s" has been updated to demonstrate this new feature.

Closes #169.