Generated files are now written in the respective build directories
The generated audiality2.pc file was in the source directory, so when
configuring multiple builds, it was overwritten. Thus obviously caused
problems when building applications, as they'd sometimes be using a .pc
file for the wrong platform!

Both audiality2.pc and audiality2.h (also generated) are now written
to the respective build directories, to avoid conflicts.

Closes #197.