Added SDL 2.x support
The build scripts now look for SDL2 first, and then SDL 1.2, unless the
new option FORCE_SDL12 is set, in which case the SDL2 check is disabled.

The SDL driver code is unchanged, as all drivers already cover the one
difference in the SDL2 audio API at this point: Buffers are not guaranteed
to be zeroed when passed to the audio callback.

The majority of changes is in a2test.c and gui.[ch], which uses various
other parts of the SDL API. The SDL2 build uses a software back buffer to
implement the dirtyrect smart screen refresh that SDL 1.2 handles mostly
by itself. Not generally the most effective design, but this application
only updates a tiny fraction of the screen area most of the time, and
also, this way most of the old SDL_Surface based rendering code could be
left mostly unchanged.

Some minor refactoring of gui.[ch] was also done. Most importantly, there
is no screen/destination pointer passed to GUI rendering calls anymore.

Closes #216.