DC generator unit with linear ramping
Added ramping DC generator with STEP and LINEAR modes. The DC generator is
(obviously) not an oscillator, but rather just outputs the 'value'
register to the audio output(s), with optional ramping.

Regardless of mode, setting the 'value' control register without ramping
always sets the audio output level instantly.

When in STEP mode, the output value switches from the current value to the
target value at 50% of the specified ramp time.

When in LINEAR mode, the output value is ramped linearily from the current
value, to the target value, over the full duration of the ramp.

(Quadratic and cubic ramping planned, but not implemented. See #156.)

Supports 1 or 2 audio outputs.

Added test program dctest.a2s, which plays various pulse, saw, triangle
and noise waves using the 'dc' unit instead of oscillators.

Closes #182.