Unit constants exported to 'units' namespace
There is now a namespace 'units', under which every voice unit that
exports constants has a namespace by the unit's name, and under that, a
namespace 'constants', which contains all constants exported by the unit.

This allows higher level scripts to use unit specific constants, without
being in a context with an instance of the unit.

Updated dctest.a2s accordingly.

Closes #221.