REFACTOR: Added a2_IsEOS() inline for checking end-of-statement tokens
Added an inline a2_IsEOS() that matches TK_EOS as well as '}', since code
looking for end-of-statement needs to accept both! This is because the
compiler needs to be able to distinguish '}', so the lexer can't return
TK_EOS in that case. (It does return TK_EOS for both newlines and ';'.)