SPAWNDR and RAMPALLR VM instructions now use the a1 operand field
The SPAWNDR and RAMPALLR instructions used to be special cases to keep the
compiler simpler, because there are related instructions that use an extra
index operand to specify target voice/register, as opposed to "all."

Now they use the a1 field, like all other VM instructions with a single
voice or register index operand. (And, the compiler code is a bit messier.
Oh, well...)

Closes #224.