Fixed segfault multiple oscillators and no panmix, or similar
When creating a voice struct with multiple units with outputs only, the
compiler wired things incorrectly, so that we ended up with a struct that
made use of scratch buffers (which it shouldn't!), but didn't actually
state that scratch buffers need to be allocated, and thus, the engine
would segfault when trying to run it.

The check for downstream inputs now honors units that can't have inputs,
instead of mistakenly assuming that any non-zero ninputs field in a struct
item means there are inputs.

When number of outputs is unspecified for a unit, the default is now to
match the channel count of the current chain, if there is one, and only
then fall back to the previous default: The minimum output count that the
unit supports. This means we can now set the output channel count for the
first unit in a chain to something other than the minimum (usually 1), and
have downstream units implicitly match that by default.

There is a new compiler exception A2_BLINDCHAIN, that's thrown as soon as
any unit tries to output to scratch buffers (that is, * or an explicit
channel count) when there are no inputs downstream, to avoid accidentally
sending output to /dev/null.

Improved DUMPSTRUCT() debug output.

Closes #239.