Forwarded imports
It's now possible to use the 'export' qualifier with 'import' statements.

In the case of a plain import, all exported symbols of the imported module
are forwarded as if they were direct exports of the current module.

If 'export' is used with 'import' <module> 'as' <name>, the imported
module is exported as a single bank symbol <name>, which serves as a
namespace containing all exports of <module>.

Namespaces now actually work with the '-p' switch to a2play, and the
separator has been changed from the former '/' (probably to avoid conflict
with filenames, which is not relevant here) to '.', to match A2S syntax.

Closes #238.