Removed broken sanity check for detaching new voices
a2_detach_or_free_handle() contained a sanity check that would warn if
someone detached a A2_TNEWVOICE handle, but since it's safe to do that if
one knows that the realtime context is not going to use it, some other
debug code would, in case of failed voice creation, change the handle type
to A2_TVOICE to disable this warning. However, that causes any further API
interaction with that handle to segfault, as the API layer assumes that an
A2_TVOICE handle always carries a valid voice pointer!

Closes #253.