REFACTOR: Control register and output naming and terminology cleanup
Reverted from "control inputs" to "control registers," as that's what they
still are at this point; read/write control registers.

Renamed A2_cout as A2_cport, as that struct is currently used for both
control outputs and control registers.

Corrected the "control outputs" comments in all unit descriptors to the
appropriate "coutputs". (Those are supposed to be field names!)