Interface separated from engine state
The new A2_interface struct replaces A2_state as the "state" argument used
throughout the API. It has a public part with call entry points for
relevant API calls, and a private implementation with further information.

The API wraps the relevant calls with inline functions, which makes it
equivalent to the old API on the source level, apart from the new type,
but the generated code for these calls is now also "DLL safe" on Windows,
allowing those calls to be used from within dynamically loaded driver and
unit code, even if Audiality 2 itself is loaded as a DLL. (See #94.)

A2_state is no longer part of the public API, not even as an opaque type.

A new call a2_AddInterface() can be used to create additional interfaces,
allowing a single engine state to have multiple independent timestamping
contexts, or interfaces with and without timestamping. This covers #204.

In order to deal with #277, a2_AddInterface() can also create A2_REALTIME
interfaces, for use in "engine context," without the usual lock-free
command FIFOs. This type of interface is also used automatically for non
realtime engine states (since there is no need for FIFOs there anyway),
eliminating the need for a2_PumpMessages() for off-line processing.

RESTRICTIONS:
  a2_Start() is not yet implemented for A2_REALTIME interfaces on
  A2_REALTIME engine states, as it is currently not safe to allocate
  new handles from the engine contexts. The recommended solution is
  to instead send messages to voices running "proxy" scripts, which
  spawn and control new voices as needed.

Closes #94.
Closes #204.
Closes #277.