Fixed some SDL/Win32 related warnings
Removed the non-const main() prototype on WIN32 hack that was still in
some of the example. Apparently, that was a bug in earlier SDL versions,
and it doesn't seem to be a problem anymore.

Removed a2_perfc_frequency initializer. It causes warnings with recent
versions of MXE, and shouldn't be needed anyway. (One isn't supposed to
use the API before initializing it, and non-WIN32 builds don't have an
initializer for their corresponding timeval struct either.)