Added VM assembly dump feature with a2play command line switch
Added API call a2_DumpCode(), which prints a VM assembly dump of the
object indicated by the provided handle. The call takes an stdio stream,
and a prefix string, for easy integration with other printouts.

a2play now accepts the '-a' switch to enable dumping of the VM assembly
code of programs with the '-x' and '-xr' switches.

Bug fix: Entry point 0 in programs now point at the main entry point, as
one would expect. (This slot was unused, as the main entry point is always
EP 0 anyway - but if it's there, it may as well be correct!)

Some minor refactoring of the VM instruction dump code.

Some whitespace fixes in compiler.c.