a2play can now play waves with the -p switch
a2play will now detect if -p is trying to play a wave, in which case it
generates and compiles and wrapper program that is then used in place of
the wave, so that it can be played like a simple "instrument" program.

Added a fallback to the A2_ROOTBANK module when a symbol isn't found in
the supposedly loaded module. This allows the -p switch and similar, to
use builtin programs, waves etc.

If no module is loaded, a2play creates a module "a2play empty module", so
that other operations will still work where applicable. (The -p switch
uses this module as a container, to manage any generated programs.)

Closes #310.