Version checking in a2_Open()
Added a2_OpenVersion(), which takes an additional argument specifying the
lowest A2 library version that's acceptable to the application.

Major and minor versions need to match. For development branches (1.9.x,
2.1.x etc), the micro versions need to match, whereas for stable branches
(2.0.x, 2.2.x etc), the library needs to be of the version requested, or
higher.

If the version check fails, a2_OpenVersion() returns NULL, and the global
error status (read with a2_LastError()) is set to A2_BADLIBVERSION.

a2_Open() is now an inline wrapper which passes the A2 header version to
the new a2_OpenVersion().

Closes #314.