Converted to SDL2 audio API with 32 bit floating point format
The SDL driver is now using the SDL2 SDL_OpenAudioDevice() interface, with
the AUDIO_F32SYS 32 bit floating point sample format.

The driver will accept the preferred sample rate and number of channels
from SDL, but buffer size is unfortunately guaranteed to be exactly what's
requested, as SDL does not yet have a flag to allow changes there.

This should eliminate the failures seen with the old API in some cases,
where we would get an unsupported sample format, just because we wanted to
allow the system's preferred sample rate and number of channels.

It also means we get full 24+ bit resolution out of A2 where possible,
which was previously only possible through the JACK driver.

Closes #340.